Corregido error al cambiar el nombre de una imagen

+++ src/app/classes/PiecesPHP/BuiltIn/Article/Controllers/ArticleController.php
@@ -707,7 +707,11 @@ class ArticleController extends AdminPanelController
 
                         //Eliminar archivo anterior
                         if (!is_null($oldFile)) {
-                            unlink($oldFile);
+
+                            if (basename($oldFile) != $nameCurrent) {
+                                unlink($oldFile);
+                            }
+
                         }
 
                         //Se elimina cualquier otro archivo
